refactor: move quick selector into its own component
