+++ src/components/layout/drawer/Pokemon.jsx
@@ -8,12 +8,7 @@ import {
   Tabs,
   List,
   ListItem,
-  ListItemText,
-  Box,
-  IconButton,
-  ButtonGroup,
   Collapse,
-  Tooltip,
   Divider,
   ListSubheader,
   Select,
@@ -22,219 +17,17 @@ import {
   InputLabel,
 } from '@mui/material'
 import { useTranslation } from 'react-i18next'
-import { VirtuosoGrid } from 'react-virtuoso'
-import Grid2 from '@mui/material/Unstable_Grid2/Grid2'
-import TuneIcon from '@mui/icons-material/Tune'
-import CheckIcon from '@mui/icons-material/Check'
-import ClearIcon from '@mui/icons-material/Clear'
 
-import { useDeepStore, useStatic, useStore } from '@hooks/useStore'
+import { useDeepStore, useStore } from '@hooks/useStore'
 import Utility from '@services/Utility'
 import { FILTER_SIZES, IV_OVERRIDES } from '@assets/constants'
 
 import { StringFilterMemo } from '../dialogs/filters/StringFilter'
 import SliderTile from '../dialogs/filters/SliderTile'
 import TabPanel from '../general/TabPanel'
-import AdvancedFilter from '../dialogs/filters/Advanced'
 import { BoolToggle, DualBoolToggle } from './BoolToggle'
-import { ItemSearchMemo } from './ItemSearch'
 import { GenderListItem } from '../dialogs/filters/Gender'
-
-function AvailableSelector() {
-  const available = useStatic((s) => s.available.pokemon)
-  const { t } = useTranslation()
-  const Icons = useStatic((s) => s.Icons)
-  const filters = useStore((s) => s.filters.pokemon)
-  const search = useStore((s) => s.searches.pokemonQuickSelect || '')
-
-  const [advanced, setAdvanced] = React.useState('global')
-  const [open, setOpen] = React.useState(false)
-
-  const items = React.useMemo(() => {
-    const lowerCase = search.toLowerCase()
-    return (
-      filters.onlyShowAvailable
-        ? available
-        : Object.keys(filters.filter).filter((key) => key !== 'global')
-    )
-      .map((key) => {
-        const [pokemon, form] = key.split('-', 2)
-        const pokemonName = t(`poke_${pokemon}`)
-        const formName = +form ? t(`form_${form}`) : ''
-
-        return {
-          key,
-          pokemonName,
-          formName,
-          url: Icons.getPokemon(pokemon, form),
-          color: filters.filter[key]?.enabled
-            ? filters.filter[key]?.all || filters.easyMode
-              ? 'success.main'
-              : 'info.main'
-            : 'error.dark',
-        }
-      })
-      .filter(
-        ({ pokemonName, formName }) =>
-          pokemonName.toLowerCase().includes(lowerCase) ||
-          formName.toLowerCase().includes(lowerCase),
-      )
-  }, [
-    filters.onlyShowAvailable,
-    filters.easyMode,
-    filters.filter,
-    available,
-    search,
-  ])
-
-  /** @param {'enable' | 'disable' | 'advanced'} action */
-  const setAll = (action) => {
-    const keys = new Set(items.map((item) => item.key))
-    useStore.setState((prev) => ({
-      filters: {
-        ...prev.filters,
-        pokemon: {
-          ...prev.filters.pokemon,
-          filter: Object.fromEntries(
-            Object.entries(prev.filters.pokemon.filter).map(([key, value]) => {
-              const enabled = action !== 'disable'
-              const all = action === 'enable'
-              return [key, keys.has(key) ? { ...value, enabled, all } : value]
-            }),
-          ),
-        },
-      },
-    }))
-  }
-
-  return (
-    <List>
-      <ItemSearchMemo field="searches.pokemonQuickSelect" />
-      <BoolToggle
-        field="filters.pokemon.onlyShowAvailable"
-        label="only_show_available"
-      />
-      <ListItem>
-        <ListItemText>{t(search ? 'set_filtered' : 'set_all')}</ListItemText>
-        <ButtonGroup variant="text" size="small" color="warning">
-          <IconButton color="success" onClick={() => setAll('enable')}>
-            <CheckIcon />
-          </IconButton>
-          <Collapse in={!filters.easyMode} orientation="horizontal">
-            <IconButton
-              color="info"
-              onClick={() => {
-                setAdvanced('global')
-                setOpen(true)
-              }}
-            >
-              <TuneIcon />
-            </IconButton>
-          </Collapse>
-          <IconButton color="error" onClick={() => setAll('disable')}>
-            <ClearIcon />
-          </IconButton>
-        </ButtonGroup>
-      </ListItem>
-      <VirtuosoGrid
-        style={{ height: 400, width: 292 }}
-        totalCount={items.length}
-        overscan={5}
-        data={items}
-        components={{
-          Item: (props) => <Grid2 xs={4} {...props} />,
-          List: React.forwardRef((props, ref) => (
-            <Grid2 {...props} container ref={ref} />
-          )),
-        }}
-        itemContent={(_, item) => {
-          const title = `${item.pokemonName}${
-            item.formName && item.formName !== t('form_29')
-              ? ` ${item.formName}`
-              : ''
-          }`
-          return (
-            <Box
-              display="flex"
-              justifyContent="center"
-              alignItems="center"
-              position="relative"
-              sx={{ aspectRatio: '1/1', outline: 'ButtonText 1px solid' }}
-              onClick={() => {
-                const filter = { ...filters.filter[item.key] }
-                if (filter.all) {
-                  filter.all = false
-                  filter.enabled = !filters.easyMode
-                } else if (filter.enabled) {
-                  filter.enabled = false
-                } else {
-                  filter.all = true
-                  filter.enabled = true
-                }
-                useStore.setState((prev) => ({
-                  filters: {
-                    ...prev.filters,
-                    pokemon: {
-                      ...prev.filters.pokemon,
-                      filter: {
-                        ...prev.filters.pokemon.filter,
-                        [item.key]: filter,
-                      },
-                    },
-                  },
-                }))
-              }}
-            >
-              <Box
-                height="100%"
-                width="100%"
-                bgcolor={item.color}
-                position="absolute"
-                top={0}
-                left={0}
-                sx={{ opacity: 0.4 }}
-              />
-              <Tooltip title={title} arrow>
-                <img
-                  alt={title}
-                  src={item.url}
-                  style={{
-                    maxHeight: 50,
-                    maxWidth: 50,
-                    zIndex: 10,
-                  }}
-                />
-              </Tooltip>
-              <Collapse in={!filters.easyMode}>
-                <IconButton
-                  size="small"
-                  sx={{ position: 'absolute', right: 0, top: 0 }}
-                  onClick={(e) => {
-                    e.stopPropagation()
-                    setAdvanced(item.key)
-                    setOpen(true)
-                  }}
-                >
-                  <TuneIcon fontSize="small" />
-                </IconButton>
-              </Collapse>
-            </Box>
-          )
-        }}
-      />
-      {!filters.easyMode && (
-        <AdvancedFilter
-          id={advanced}
-          category="pokemon"
-          open={open}
-          setOpen={setOpen}
-        />
-      )}
-    </List>
-  )
-}
-
-const MemoAvailableSelector = React.memo(AvailableSelector, () => true)
+import { MemoAvailableSelector } from './ItemSelector'
 
 export default function WithSliders({ category, context }) {
   const userSettings = useStore((s) => s.userSettings[category])
@@ -371,7 +164,7 @@ export default function WithSliders({ category, context }) {
             </TabPanel>
           ))}
           <TabPanel value={openTab} index={2} disablePadding>
-            <MemoAvailableSelector />
+            <MemoAvailableSelector category="pokemon" />
           </TabPanel>
         </>
       )}
